Fixed deadlock in GameScheduler
When putting the game back on the queue it needs to be done in a go
routine or it will block the select.
Added the debug flag, setting it to false will disable the log output.
Added a few logging statements.
Removed the launcher from the game list sent to clients.