Started working on the new back end desgin
New clients should now be asked to identify themselves when connecting.
A identify response should be on the form:
{"action":"identify","data":"name"}
Sending a identify message from a new connection will overwrite the old
one.
This is so the back end can resume a broken connection.
Sending a new identify message while connected will rename the client.