The game and controllers should now be able to pass messages to each other
Added To and From fields to the message structs.
Added the Raw field to outgoing messages.
This field will be used when the action is pass through as we don't
know (or want to know) what Data contains.
Changed sendToGame channel to a generic send channel.
Added constants for actions identify and passthrough.