Redesigned how connections work
Added a new struct, request:
When registering a new client we'll be sending a request to the hub.
If a client with the same id as in the request already exists we can
return the old connection and resume it.
We are basically just swapping the old ws for the new one.
If the id is free, we make a new connection.
I've added two new fields to the connection struct, active and timeout.
When active is false it means that the current ws connection is
inactive.
If we get a new connection claiming to be the inactive id, we can set
active back to true.
By doing so we stop the old connection from removing the id from
clients.
That is what timeout is for, it waits a duration after the connection
becomes inactive, then removes itself if active is still false.
Unregister now only needs to take the id instead of the whole
connection.
Removed hub_test.go, tests need to be rewritten.
Removed the rename action, we won't allow renaming for now.
Removed the connections field from the hub, we don't need it anymore.