Fixed a some concurrency problems
Messages sent while a connection is broken are now stored and will be
sent when the connection resumes.
Registrations to the hub are now done with a registration struct.
The ok channel should be used to return the result of the registration
and must always be read or the hub will deadlock.
The connection should now correctly send a close message and close the
ws connection when a error is encountered.
Resuming a connection should now make the old connection return
immediately.
hub.go now have 100% test coverage.