Fixed multiple data races
Put a lock on connection.active as it being both read and written by
multiple go routines.
Implemented the functions setActive and isActive which handles the
locking behind the scenes.
Removed the closeWs function and restored how the close message was
originally sent.
The problem was multiple calls to ws.WriteMessage.
Added connection.stop a channel for signaling the connection to
terminate early.