Modified hub.js to handle identification
hub.js will now pick a random id for you.
We will deal with collisions on ids later.
The id is stored locally and remembered for further use, unless the user is using private browsing or similar.
addMessageHandler now parses the json before it's sent to the callback.
The callback function is called with the string "identified" when the
identify message has been sent.