Clients now receieve a redirect when added/resumed
Also when a game is started, a redirect message is broadcasted to all
clients.
The only time a redirect does not get sent is when a game is quit and
the gamescheduler directly starts the next game.
I hope to fix this in #24 by allowing callbacks to be registered for
game start events.
Closes #15.