Unfinished work on the controller for pong
I know that the message check on line 31:42 will have to change as I figured that the id might already been set if pong is launched through the controller.
How the controller removes the buttons (94:99) when they are set will also have to be changed some as you need to revive the buttons if they become available.