Added package description for new launcher
The new launcher will be using nw.js.
nw.js allows us to write native apps in html and javascript.
This should allow us to reuse much of the js we have already written.
Designing the application should also be a lot easier.
Using nw.js we will need to generate a executable for each platform we will be
supporting.
This means we need to expand game.json and the back end to include exec commands for different
platforms.